pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-community>=0.3.27",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
    "python-multipart>=0.0.9",
    "pypdf>=5.1.0",
    "pdfplumber>=0.11.0",
    "watchdog>=6.0.0",
]
